# web-app/openapi.yaml
openapi: 3.1.0
info:
  title: FlowShield API
  version: 1.9.5
  description: >
    FlowShield productivity tracking API.
    All endpoints except auth require a Bearer JWT token
    obtained from /api/auth/login or /api/auth/google.
  contact:
    name: FlowShield
    url: https://flowshield.app
servers:
  - url: https://flowshield.app
    description: Production
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/logout-all:
    post:
      tags: [Auth]
      summary: Log out of all devices
      description: Increments the user's tokenVersion, revoking every previously issued JWT.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: All tokens revoked
        "401":
          description: Missing or invalid token
        "429":
          description: Rate limited (5/hour per user)
...